chore: Update getPlatformUserBalances response type - Separate balances from user endpoints on example-node-api-client
